requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
boto3==1.35.35
botocore==1.35.35
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1
django-celery-results==2.5.1
django-tinymce==4.1.0
dnspython==2.7.0
eventlet==0.37.0
flower==2.0.1
greenlet==3.1.1
humanize==4.10.0
jmespath==1.0.1
kombu==5.4.2
pillow==10.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
